--- Apester/Mustache/Apester.mustache
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1"/>
-</head>
-<body>
-
-    <div class="apester-media" data-media-id="{{{ mediaId }}}" height="493"></div>
-
-    <script async src="https://static.apester.com/js/sdk/v2.0/apester-javascript-sdk.min.js"></script>
-
-</body>
-</html>
