+++ Apester/View/APEStripCollectionViewCell.swift
