--- ApesterKit.xcworkspace/xcshareddata/ApesterKit.xcscmblueprint
@@ -1,44 +1,0 @@
-{
-  "DVTSourceControlWorkspaceBlueprintPrimaryRemoteRepositoryKey" : "307FCDC9323289AE9ECD68ADB599712096860267",
-  "DVTSourceControlWorkspaceBlueprintWorkingCopyRepositoryLocationsKey" : {
-
-  },
-  "DVTSourceControlWorkspaceBlueprintWorkingCopyStatesKey" : {
-    "307FCDC9323289AE9ECD68ADB599712096860267" : 9223372036854775807,
-    "95438028B10BBB846574013D29F154A00556A9D1" : 9223372036854775807,
-    "D0725CAC6FF2D66F2C83C2C48DC12106D42DAA64" : 9223372036854775807,
-    "E084C86B03F81D63323C9E7510697EA528A758C7" : 9223372036854775807
-  },
-  "DVTSourceControlWorkspaceBlueprintIdentifierKey" : "58655E50-7052-4E14-997C-F87FF22D53DE",
-  "DVTSourceControlWorkspaceBlueprintWorkingCopyPathsKey" : {
-    "307FCDC9323289AE9ECD68ADB599712096860267" : "ios-sdk\/",
-    "95438028B10BBB846574013D29F154A00556A9D1" : "ios-sdk\/Carthage\/Checkouts\/Nimble\/",
-    "D0725CAC6FF2D66F2C83C2C48DC12106D42DAA64" : "ios-sdk\/Carthage\/Checkouts\/Quick\/",
-    "E084C86B03F81D63323C9E7510697EA528A758C7" : "ios-sdk\/Carthage\/Checkouts\/xcconfigs\/"
-  },
-  "DVTSourceControlWorkspaceBlueprintNameKey" : "ApesterKit",
-  "DVTSourceControlWorkspaceBlueprintVersion" : 204,
-  "DVTSourceControlWorkspaceBlueprintRelativePathToProjectKey" : "ApesterKit.xcworkspace",
-  "DVTSourceControlWorkspaceBlueprintRemoteRepositoriesKey" : [
-    {
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryURLKey" : "github.com:Qmerce\/ios-sdk.git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositorySystemKey" : "com.apple.dt.Xcode.sourcecontrol.Git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryIdentifierKey" : "307FCDC9323289AE9ECD68ADB599712096860267"
-    },
-    {
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryURLKey" : "https:\/\/github.com\/Quick\/Nimble.git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositorySystemKey" : "com.apple.dt.Xcode.sourcecontrol.Git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryIdentifierKey" : "95438028B10BBB846574013D29F154A00556A9D1"
-    },
-    {
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryURLKey" : "https:\/\/github.com\/Quick\/Quick.git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositorySystemKey" : "com.apple.dt.Xcode.sourcecontrol.Git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryIdentifierKey" : "D0725CAC6FF2D66F2C83C2C48DC12106D42DAA64"
-    },
-    {
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryURLKey" : "https:\/\/github.com\/jspahrsummers\/xcconfigs.git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositorySystemKey" : "com.apple.dt.Xcode.sourcecontrol.Git",
-      "DVTSourceControlWorkspaceBlueprintRemoteRepositoryIdentifierKey" : "E084C86B03F81D63323C9E7510697EA528A758C7"
-    }
-  ]
-}
