--- Cartfile.private
@@ -1,4 +1,0 @@
-github "jspahrsummers/xcconfigs" "master"
-github "Quick/Quick"
-github "Quick/Nimble"
-github "groue/GRMustache.swift"
