--- docs/docsets/ApesterKit.docset/Contents/Info.plist
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-  <dict>
-    <key>CFBundleIdentifier</key>
-      <string>com.jazzy.apesterkit</string>
-    <key>CFBundleName</key>
-      <string>ApesterKit</string>
-    <key>DocSetPlatformFamily</key>
-      <string>apesterkit</string>
-    <key>isDashDocset</key>
-      <true/>
-    <key>dashIndexFilePath</key>
-      <string>index.html</string>
-    <key>isJavaScriptEnabled</key>
-      <true/>
-    <key>DashDocSetFamily</key>
-      <string>dashtoc</string>
-  </dict>
-</plist>
