--- docs/docsets/ApesterKit.docset/Contents/Resources/Documents/search.json
@@ -1,1 +1,0 @@
-{"Protocols.html":{"name":"APEWebViewProtocol","abstract":"<p>APEWebViewProtocol where UIWebView &amp;&amp; WKWebView confirms to it</p>"},"Enums/APEResult.html#/s:10ApesterKit9APEResultO7successACyxGxcAEmlF":{"name":"success","abstract":"<p>success case with success param</p>","parent_name":"APEResult"},"Enums/APEResult.html#/s:10ApesterKit9APEResultO7failureACyxGSScAEmlF":{"name":"failure","abstract":"<p>failure case with error description</p>","parent_name":"APEResult"},"Enums/APEResult.html":{"name":"APEResult","abstract":"<p>APEWebViewServiceResult enum</p>"},"Classes/APEWebViewService.html#/s:10ApesterKit17APEWebViewServiceC6sharedACvZ":{"name":"shared","abstract":"<p>APEWebViewService shared instance</p>","parent_name":"APEWebViewService"},"Classes/APEWebViewService.html#/c:@CM@ApesterKit@objc(cs)APEWebViewService(im)userContentController:didReceiveScriptMessage:":{"name":"userContentController(_:didReceive:)","abstract":"<p>Undocumented</p>","parent_name":"APEWebViewService"},"Classes/APEWebViewService.html#/s:10ApesterKit17APEWebViewServiceC8registerySo6BundleC6bundle_AA0cD8Protocol_p03webD0yAA9APEResultOy12CoreGraphics7CGFloatVGcSg17unitHeightHandleryAKySbGcSg010completionQ0tF":{"name":"register(bundle:webView:unitHeightHandler:completionHandler:)","abstract":"<p>call register(bundle:webview:unitHeightHandler:completionHandler:) function from viewDidLoad</p>","parent_name":"APEWebViewService"},"Classes/APEWebViewService.html#/s:10ApesterKit17APEWebViewServiceC12didStartLoadyAA0cD8Protocol_p03webD0_yAA9APEResultOySbGcSg17completionHandlertF":{"name":"didStartLoad(webView:completionHandler:)","abstract":"<p>call didStartLoad(webView:) function once the webview delegate didStartLoad triggered,","parent_name":"APEWebViewService"},"Classes/APEWebViewService.html#/s:10ApesterKit17APEWebViewServiceC13didFinishLoadyAA0cD8Protocol_p03webD0_yAA9APEResultOySbGcSg17completionHandlertF":{"name":"didFinishLoad(webView:completionHandler:)","abstract":"<p>call didFinishLoad(webView:) function once the webview delegate didFinishLoad triggered,","parent_name":"APEWebViewService"},"Classes/APEWebViewService.html":{"name":"APEWebViewService","abstract":"<p>APEWebViewService provides a light-weight framework that loads Apester Unit in a webView</p>"},"Classes.html":{"name":"Classes","abstract":"<p>The following classes are available globally.</p>"},"Enums.html":{"name":"Enumerations","abstract":"<p>The following enumerations are available globally.</p>"},"Protocols.html":{"name":"Protocols","abstract":"<p>The following protocols are available globally.</p>"}}
