--- docs/undocumented.json
@@ -1,6 +1,0 @@
-{
-  "warnings": [
-
-  ],
-  "source_directory": "/private/var/www/ios-sdk"
-}
