--- scripts/bootstrap
@@ -1,1 +1,0 @@
-carthage update --no-use-binaries --use-submodules
